+++ pytsite/theme/_eh.py
@@ -2,7 +2,8 @@
 """
 from os import path as _path
 from shutil import move as _move
-from pytsite import http as _http, router as _router, settings as _settings, reg as _reg, console as _console
+from pytsite import http as _http, router as _router, settings as _settings, reg as _reg, console as _console, \
+    file as _file, metatag as _metatag
 from . import _api
 
 __author__ = 'Alexander Shepetko'
@@ -24,6 +25,17 @@ def router_dispatch():
         _api.set_current(request_cookies['PYTSITE_THEME'])
     elif theme_setting and _api.is_registered(theme_setting):
         _api.set_current(theme_setting)
+
+    # Set favicon URL
+    favicon_fid = _settings.get('theme.favicon_fid')
+    if favicon_fid:
+        try:
+            f = _file.get(favicon_fid)
+            _metatag.rm('link', rel='icon')
+            _metatag.t_set('link', rel='icon', type=f.mime, href=f.get_url(width=50, height=50))
+        except _file.error.FileNotFound:
+            pass
+
 
 
 def router_response(response: _http.response.Response):